Rename and update command pattern
